Refactoring with separating of plugins and resolve logic with filters
