+++ src/plugins/unpkg-path-plugin.ts
@@ -1,76 +1,29 @@
 import * as esbuild from 'esbuild-wasm';
-import axios from 'axios';
-import localForage from 'localforage';
-
-const fileCache = localForage.createInstance({
-  name: 'filecache'
-});
-
-(async () => {
-  await fileCache.setItem('color', 'red');
-
-  const color = await fileCache.getItem('color');
-  console.log(color);
-})();
 
 export const unpkgPathPlugin = () => {
   return {
     name: 'unpkg-path-plugin',
     setup(build: esbuild.PluginBuild) {
-      build.onResolve({ filter: /.*/ }, async (args: any) => {
-        console.log('onResolve', args);
-        if (args.path === 'index.js') {
-          return { path: args.path, namespace: 'a' };
-        }
-        if (args.path.includes('./') || args.path.includes('../')) {
-          return {
-            namespace: 'a',
-            path: new URL(
-              args.path,
-              'https://unpkg.com' + args.resolveDir + '/'
-            ).href
-          };
-        }
+      // Handle root entry file of index.js
+      build.onResolve({ filter: /(^index\.js$)/ }, () => {
+        return { path: 'index.js', namespace: 'a' };
+      });
+
+      // handle relative paths in a module
+      build.onResolve({ filter: /^\.+\// }, async (args: any) => {
         return {
           namespace: 'a',
-          path: `https://unpkg.com/${args.path}`
+          path: new URL(args.path, 'https://unpkg.com' + args.resolveDir + '/')
+            .href
         };
       });
 
-      build.onLoad({ filter: /.*/ }, async (args: any) => {
-        console.log('onLoad', args);
-
-        if (args.path === 'index.js') {
-          return {
-            loader: 'jsx',
-            contents: `
-                  import React from 'react';
-                  console.log(react);
-                `
-          };
-        }
-
-        // Chcek if we have the fetched file and if it is in the cache
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
-          args.path
-        );
-
-        // if it is return it immediately
-        if (cachedResult) {
-          return cachedResult;
-        }
-
-        // Build the object
-        const { data, request } = await axios.get(args.path);
-
-        const result: esbuild.OnLoadResult = {
-          loader: 'jsx',
-          contents: data,
-          resolveDir: new URL('./', request.responseURL).pathname
+      // Handle main file in a module
+      build.onResolve({ filter: /.*/ }, async (args: any) => {
+        return {
+          namespace: 'a',
+          path: `http://unpkg.com/${args.path}`
         };
-        // store response in cache
-        await fileCache.setItem(args.path, result);
-        return result;
       });
     }
   };
